+++ subsys/portability/cmsis_rtos_v2/mempool.c
@@ -5,6 +5,7 @@
  */
 
 #include <zephyr/kernel.h>
+#include <zephyr/portability/cmsis_types.h>
 #include <string.h>
 #include "wrapper.h"
